Add files via upload
Optimal use of Vehicle-to-grid technology to modify load profile of distribution system 
Abstract--Managing the use of electric vehicles and power injections from the batteries of these vehicles poses the issue of controlling the charge and discharge of electric vehicles as an attractive research field. Charging a large number of electric vehicles batteries will, if not controlled, have a negative effect on the distribution system. By adopting optimal planning for the use of electric vehicles, parking of this type of vehicle can operate in two roles of load or source of energy for the distribution network. On the other hand, the injectable power of the parking lots can cause the displacement of the peak times of the system. In this thesis, the effect of scheduling the status of charging and discharging of electric vehicles on load characteristic enhancement. On the other hand, the power of parking of electric vehicles is uncertain and probable. Therefore, the calculation takes into account the probabilistic model of the load and the production of each type of parking lot. Optimization is based on a hybrid method including metaheuristic JAYA algorithm and Monte Carlo. Therefore, in this research, the first step of the probabilistic model is extracted from the behavior of the owners of cars. Then, using this behavior and the demands of the owners, the scheduling of charging and discharging the electrical vehicle is determined so that the daily load variance reduces and the network load characteristic becomes smooth.